study-hub/requirements.txt
boto3==1.40.57
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.6
django-autoslug==1.9.9
django-crispy-forms==2.4
django-environ==0.12.0
django-storages==1.14.6
GitPython==3.1.45
gunicorn==23.0.0
mysqlclient==2.2.7
pillow==11.3.0
whitenoise==6.10.0
